Changes to file structure
