+++ src/View/SubMenus/Attributes/Atributos.h
@@ -1,0 +1,13 @@
+#ifndef ATRIBUTOS_H
+#define ATRIBUTOS_H
+
+#include "../../terminal.h"
+#include "../../Utils/io.h"
+
+TerminalApp getSubmenuAtributos();
+
+
+
+
+
+#endif //ATRIBUTOS_H
