+++ src/View/SubMenus/Category/Categorias.h
@@ -1,0 +1,12 @@
+#ifndef CATEGORIAS_H
+#define CATEGORIAS_H
+
+#include "../../terminal.h"
+#include "../../Utils/io.h"
+
+
+TerminalApp getSubmenuCategorias();
+
+
+
+#endif //CATEGORIAS_H
