+++ src/View/SubMenus/Products/Productos.h
@@ -1,0 +1,10 @@
+#ifndef PRODUCTOS_H
+#define PRODUCTOS_H
+
+#include "../../terminal.h"
+#include "../../Utils/io.h"
+
+TerminalApp getSubmenuProductos();
+
+
+#endif //PRODUCTOS_H
