+++ src/View/SubMenus/Search/Busqueda.h
@@ -1,0 +1,9 @@
+#ifndef BUSQUEDA_H
+#define BUSQUEDA_H
+
+#include "../../terminal.h"
+#include "../../Utils/io.h"
+
+TerminalApp getSubmenuBusqueda();
+
+#endif //BUSQUEDA_H
