+++ src/View/SubMenus/Variants/Variantes.h
@@ -1,0 +1,10 @@
+#ifndef VARIANTES_H
+#define VARIANTES_H
+
+#include "../../terminal.h"
+#include "../../Utils/io.h"
+
+TerminalApp getSubmenuVariantes();
+
+
+#endif //VARIANTES_H
