+++ src/View/terminal.cpp
@@ -1,4 +1,4 @@
-#include "terminal.h"
+#include "./terminal.h"
 #include <limits>
 
 void TerminalApp::execute() {
