+++ src/View/terminal.h
@@ -7,7 +7,7 @@
 #include <memory>
 #include <functional>
 #include <sstream>
-#include "io.h"
+#include "./Utils/io.h"
 
 class TerminalOperation {
 public:
