+++ src/View/terminal_implementations.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "terminal.h"
+#include "../View/Utils/io.h"
+
+TerminalApp GetAppInstance();
+
