--- src/terminal_implementations.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#include "terminal.h"
-#include "io.h" 
-
-TerminalApp getSubmenuProductos();
-TerminalApp getSubmenuCategorias();
-TerminalApp getSubmenuBusqueda();
-TerminalApp GetAppInstance();
-
